# backend/internal/apispec/openapi.yaml
openapi: 3.0.3
info:
  title: XCreativs Web Platform API
  version: "1.0"
  description: >
    REST API for the XCreativs platform. Bearer (JWT) auth is offered globally via the
    Authorize button; public endpoints ignore the token. Scoped API keys use the
    X-API-Key header. Generated from the live router — do not edit by hand;
    run `go run ./cmd/genopenapi > api/openapi.yaml`.
servers:
  - url: http://localhost:8081
    description: Local
  - url: /
    description: Same origin
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
    apiKeyAuth: { type: apiKey, in: header, name: X-API-Key }
  schemas:
    Error:
      type: object
      properties:
        error: { type: string }
  responses:
    Error:
      description: Error
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
security:
  - bearerAuth: []
paths:
...
  "/api/v1/audio-briefs":
    get:
      tags: [Public]
      summary: GET /api/v1/audio-briefs
      operationId: get_api_v1_audio_briefs
      responses:
        '200': { description: Success }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        default: { $ref: '#/components/responses/Error' }
...
  "/api/v1/feed/insights":
    get:
      tags: [Public]
      summary: GET /api/v1/feed/insights
      operationId: get_api_v1_feed_insights
      responses:
        '200': { description: Success }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        default: { $ref: '#/components/responses/Error' }
...
  "/api/v1/insights":
    get:
      tags: [Public]
      summary: GET /api/v1/insights
      operationId: get_api_v1_insights
      responses:
        '200': { description: Success }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        default: { $ref: '#/components/responses/Error' }
...
  "/api/v1/invoices/{id}":
    patch:
      tags: [Portal]
      summary: PATCH /api/v1/invoices/{id}
      operationId: patch_api_v1_invoices_id
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: false
        content:
          application/json:
            schema: { type: object }
      responses:
        '200': { description: Success }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        default: { $ref: '#/components/responses/Error' }
...
  "/api/v1/portal/engagements/{id}/invoices":
    get:
      tags: [Portal]
      summary: GET /api/v1/portal/engagements/{id}/invoices
      operationId: get_api_v1_portal_engagements_id_invoices
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200': { description: Success }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        default: { $ref: '#/components/responses/Error' }
    post:
      tags: [Portal]
      summary: POST /api/v1/portal/engagements/{id}/invoices
      operationId: post_api_v1_portal_engagements_id_invoices
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: false
        content:
          application/json:
            schema: { type: object }
      responses:
        '200': { description: Success }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        default: { $ref: '#/components/responses/Error' }
...